Fixed qulice issues
