+++ eo-archetype/src/main/resources/META-INF/maven/archetype-metadata.xml
@@ -1,4 +1,8 @@
-<archetype-descriptor name="eo-archetype">
+<?xml version="1.0" encoding="UTF-8"?>
+<archetype-descriptor xmlns="http://maven.apache.org/plugins/maven-archetype-plugin/archetype-descriptor/1.0.0"
+                      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                      xsi:schemaLocation="http://maven.apache.org/plugins/maven-archetype-plugin/archetype-descriptor/1.0.0 http://maven.apache.org/xsd/archetype-descriptor-1.0.0.xsd"
+                      name="eo-archetype">
     <fileSets>
         <fileSet filtered="true" encoding="UTF-8">
             <directory>src/main/eo</directory>
